Create a custom @RoleAuthorization annotation
I've created that annotation to distinguish cases when request
has been sent by unauthorized user or authorized user. Spring
Security in both cases throws AccessDeniedException. I need to
distinguish it. The annotation throws AccessDeniedException
when user is unauthorized and custom ForbiddenAccessException
when user is authorized but has no permission to get data.
It's more comfortable to manage that both cases.