Update Authentication service
Access and refresh tokens are stored in th session storage.
User is authenticated until the both tokens are valid.
When the access token expires, a silent refresh will happen.
I've created a timer that, based on the access token's expiration
time, makes a request to renew the token 1 minute before it
expires.
When the refresh token expires, user will be logged out for
security reasons.
Logout method sends a request to server to make user tokens
invalid.