Create an HTTP interceptor
The http interceptor adds an authorization header with a token
to each request. It adds the refresh token when the /refresh-token
endpoint is executed; in other cases it adds the access token.
It also intercepts 401 error - if the access token is expired,
the interceptor attempts to refresh it and resend the request.