[FEATURE][F-53] Implement the notifications page (#67)
* Add WebSocket services for backend and frontend

Added dependency to spring-boot-starter-websocket in the backend's
POM file. Implemented WebSocket service in the backend.

Created a WebSocket service on the frontend to communicate with the
backend service.

This setup enables real-time communication between the frontend and
backend using WebSockets.

* Add Notification entity and setup in backend and frontend

Created Notification entity with necessary fields, replacing the
previous simple class. Added NotificationType enum containing types
of notifications, each with a reason and content.

Created a Liquibase changelog to create the notification table for
the new entity. Added test data in .sql files and updated master.xml
to include these files for the test and development profiles.

Added NotificationRepository using JPA for data access.

Implemented a corresponding Notification entity on the frontend to
match the backend changes.

* Implement NotificationService in the backend

Created NotificationService with necessary methods and a DTO
(NotificationDto) corresponding to the Notification entity. Added a
mapper for conversion between entity and DTO.

Introduced a custom exception (NotificationNotFoundException) and
updated the exception handler to catch it when a notification is
not found.

Utilized the Strategy pattern for sending messages via
NotificationStrategy interface. Implemented three strategies:
system notifications, SMS notifications, and Email notifications.

Updated classes and message content for user actions related to
sending system, email, and SMS notifications. Removed obsolete
classes that became unnecessary after introducing the strategy
pattern.

* Send realtime notifications after user actions

Injected NotificationService into ReservationService and
LendingService. Now, appropriate notifications are sent to the
user after executing the main methods of these services that
correspond to user actions.

* Add NotificationController with CRUD endpoints + tests

Added NotificationController with endpoints for the following actions:
  -  Retrieve all notifications for a user
  -  Retrieve a single notification
  -  Mark a notification as 'read'
  -  Delete a single notification
  -  Delete multiple notifications

Updated CustomSecurityConfig to include the DELETE method in allowed
CORS methods. Added integration tests for NotificationController to
ensure proper functionality.

* Add NotificationService in the frontend

Added NotificationService on the frontend to handle requests to the
backend. Created a class to represent the Pageable object returned
by the backend for the endpoint fetching a collection of
notifications. This ensures proper handling of paginated notification
data on the frontend.

* Implement real-time notification list

Implemented NotificationComponent with an HTML template and CSS,
displaying a table of notifications updated in real-time. Added
buttons for selecting all notifications at once and for deleting
selected notifications.

Notification titles are linked to future detail views for each
notification, preparing for detailed notification display
functionality.

* Add NotificationDetailComponent with detailed view

Created a new NotificationDetailComponent to display the details of a
single notification. Designed the HTML template and added styles. The
page shows information such as the notification's creation date,
content, and the title of the related book, with a link to the book's
page.

Additionally, added buttons for deleting the notification and a "Back"
button for navigation.

* Standalone PaginationComponent and code refactor

 - PaginationComponent is now standalone, requiring only the 'page'
   attribute
 - Removed unnecessary BookItemListComponent
 - Refactored BookListComponent to delegate operations to BookService
 - Moved logic from BookListComponent to BookService, providing
   observables
 - Moved model files from 'shared' to 'models' folder
 - Replaced ListComponent interface with Pageable, added Sortable
   interface and helper classes
 - Replaced Pair class with Language class to match the returned object

* Add pagination to notification page

- Backend: Added a repository method to fetch notifications in
  descending order (by default) and updated test data + tests.
- Frontend: Improved methods for removing notifications and added
  pagination to notifications list in HTML.

* Fix 'Subject' column width to 100%